requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
SQLAlchemy==2.0.21
bcrypt==4.0.1
python-dotenv==1.0.0
Jinja2==3.1.2
email-validator==2.0.0
